Further refactoring of the tests.
