+++ tests/test_flask/test_views/test_reset_view.py
@@ -1,0 +1,35 @@
+"""Test the reset view using flask's test_client()."""
+from run import db
+
+
+def test_dropped_db(client):
+    """Check behavior when the db has no tables."""
+    db.drop_all()
+    assert b'Sorry' in client.get('/').data
+    assert "404" in client.get('/poem/Flea').status
+    # /reset route causes server error when db has no tables
+
+
+def test_reset_route(client_empty_db):
+    """Check the /reset route."""
+    # The database has tables in it
+    assert "poems" in db.engine.table_names()
+    # The database has no poems in it
+    assert b'Sorry' in client_empty_db.get('/').data
+    assert "404" in client_empty_db.get('/poem/Flea').status
+
+    # /reset route injects sample poems into database
+    client_empty_db.get('/reset')
+    assert b'Flea' in client_empty_db.get('/').data
+    response = client_empty_db.get('/poem/Flea')
+    assert b'Flea' in response.data
+    assert "200" in response.status
+    assert "404" in client_empty_db.get('/poem/foo').status
+
+    # Makes sure doing it twice doesn't break anything.
+    client_empty_db.get('/reset')
+    assert b'Flea' in client_empty_db.get('/').data
+    response = client_empty_db.get('/poem/Flea')
+    assert b'Flea' in response.data
+    assert "200" in response.status
+    assert "404" in client_empty_db.get('/poem/foo').status
